Add missing dependency for manager cloud-init
A dependency on manager cloud-init completion was missing. Since we
have to use a null_resource to wait for manager cloud-init completion
due to the (potential) use of managed disks which cannot be created
simultaneously with the manager (or workers) in terraform, the normal
dependency mechanism doesn't work. Since the config-files module
depends on the results of manager cloud-init completion we need the
explicit dependency added.